# docs/openapi.yaml
openapi: "3.1.0"
info:
  title: FlowLink API
  version: "1.0.0"
  description: >-
    FlowLink compliance, identity, and invoicing API for autonomous agent
    payments. Provides sanctions screening, KYA credential issuance, invoice
    lifecycle management, webhook delivery, and platform analytics.
  contact:
    name: FlowLink
    url: https://flowlink.io
  license:
    name: MIT
servers:
  - url: "{baseUrl}/v1"
    description: FlowLink API v1
    variables:
      baseUrl:
        default: "http://localhost:3001"
        description: Base URL of the API server
security:
  - ApiKeyHeader: []
  - BearerAuth: []
...
components:
  securitySchemes:
    ApiKeyHeader:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key passed via X-API-Key header
    BearerAuth:
      type: http
      scheme: bearer
      description: API key or JWT passed via Authorization Bearer header
  schemas:
    ErrorResponse:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          const: false
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Machine-readable error code
            message:
              type: string
              description: Human-readable error message
            details:
              description: Validation error details (array of { path, message })
...
    VerifyAgentRequest:
      type: object
      required: [agentId]
      properties:
        agentId:
          type: string
          description: Agent identifier (DID or registry ID)
        registryAddress:
          type: string
          description: Optional ERC-8004 registry contract address
        chain:
          type: string
          default: "eip155:8453"
          description: CAIP-2 chain identifier
    VerifyAgentResponse:
      type: object
      properties:
        success:
          type: boolean
          const: true
        data:
          type: object
          properties:
            verified:
              type: boolean
            trustScore:
              type: integer
              minimum: 0
              maximum: 100
            agentMetadata:
              type: object
              nullable: true
              properties:
                name:
                  type: string
                type:
                  type: string
                  enum: [autonomous, semi-autonomous, human-supervised]
                operator:
                  type: string
                registeredAt:
                  type: string
                  format: date-time
                walletAddress:
                  type: string
            operatorStatus:
              type: object
              properties:
                sanctionsCleared:
                  type: boolean
                kycVerified:
                  type: boolean
            delegationScope:
              type: object
    ControllingEntity:
      type: object
      required: [name, kybVerified]
      properties:
        name:
          type: string
        lei:
          type: string
        did:
          type: string
        kybVerified:
          type: boolean
    DelegationScope:
      type: object
      required: [maxTransactionValue, expiresAt]
      properties:
        maxTransactionValue:
          type: number
          minimum: 0
        dailyLimit:
          type: number
          minimum: 0
        allowedCounterparties:
          type: array
          items:
            type: string
        blockedJurisdictions:
          type: array
          items:
            type: string
        allowedChains:
          type: array
          items:
            type: string
        allowedCurrencies:
          type: array
          items:
            type: string
        expiresAt:
          type: string
          format: date-time
    IssueKYARequest:
      type: object
      required:
        - agentDid
        - agentType
        - controllingEntity
        - walletAddress
        - delegationScope
      properties:
        agentDid:
          type: string
        agentType:
          type: string
          enum: [autonomous, semi-autonomous, human-supervised]
        controllingEntity:
          $ref: "#/components/schemas/ControllingEntity"
        walletAddress:
          type: string
        delegationScope:
          $ref: "#/components/schemas/DelegationScope"
        erc8004RegistryAddress:
          type: string
        erc8004TokenId:
          type: string
...
  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
paths:
...
  /identity/verify:
    post:
      tags: [Identity]
      summary: Verify agent KYA identity
      description: >-
        Verifies an agent's KYA credential and returns trust score, metadata,
        and delegation scope.
      operationId: verifyAgent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyAgentRequest"
      responses:
        "200":
          description: Verification result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerifyAgentResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /identity/{agentId}:
    get:
      tags: [Identity]
      summary: Get agent identity info
      operationId: getAgentIdentity
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: string
          description: Agent DID
      responses:
        "200":
          description: Agent identity
        "404":
          $ref: "#/components/responses/NotFound"
  /identity/kya/issue:
    post:
      tags: [Identity]
      summary: Issue KYA credential
      description: >-
        Issues or re-issues a KYA (Know-Your-Agent) verifiable credential.
        Creates or updates the agent record and returns a W3C VC-shaped
        credential.
      operationId: issueKYA
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IssueKYARequest"
      responses:
        "201":
          description: KYA credential issued
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
...